fix style duplication
