+++ package.json
@@ -48,7 +48,7 @@
     "@storybook/react": "^5.2.8",
     "@types/jest": "^24.9.0",
     "@types/react": "^16.9.19",
-    "@types/styled-components": "^4.4.0",
+    "@types/styled-components": "^5.0.1",
     "@types/styled-system": "^5.1.4",
     "@typescript-eslint/eslint-plugin": "^2.19.0",
     "@typescript-eslint/parser": "^2.19.0",
@@ -65,7 +65,7 @@
     "eslint-plugin-standard": "^4.0.1",
     "jest": "^24.9.0",
     "react": "^16.12.0",
-    "styled-components": "^4.4.1",
+    "styled-components": "^5.0.1",
     "ts-jest": "^24.3.0",
     "ts-loader": "^6.2.1",
     "ts-node": "^8.6.2",
