+++ src/components/AspectRatio.tsx
@@ -1,15 +1,16 @@
 import React, { FC } from 'react';
 import styled from 'styled-components';
 
-import { boxBase, BoxProps } from './Box';
+import { boxBase, BoxBaseProps } from '../bases';
+import { compose } from '../utils/baseStyle';
 import Flex from './Flex';
 
 export type AspectRatioProps = {
   ratio: number;
-} & BoxProps;
+} & BoxBaseProps;
 
-const AspectRatioStyle = styled.div<BoxProps>`
-  ${p => boxBase(p)}
+const AspectRatioStyle = styled.div<BoxBaseProps>`
+  ${compose(boxBase)}
 
   & > div {
     width: 100%;
@@ -22,24 +23,26 @@ const AspectRatioStyle = styled.div<BoxProps>`
   }
 `;
 
-const AspectRatio: FC<AspectRatioProps> = ({ ratio, children, ...props }) => (
-  <AspectRatioStyle
-    width="100%"
-    height={0}
-    pb={ratio * 100 + '%'}
-    position="relative"
-    {...props}
-  >
-    <Flex
-      position="absolute"
-      alignItems="center"
-      justifyContent="center"
-      left={0}
-      top={0}
+const AspectRatio: FC<AspectRatioProps> = ({ ratio, children, ...props }) => {
+  return (
+    <AspectRatioStyle
+      width="100%"
+      height={0}
+      pb={ratio * 100 + '%'}
+      position="relative"
+      {...props}
     >
-      {children}
-    </Flex>
-  </AspectRatioStyle>
-);
+      <Flex
+        position="absolute"
+        alignItems="center"
+        justifyContent="center"
+        left={0}
+        top={0}
+      >
+        {children}
+      </Flex>
+    </AspectRatioStyle>
+  );
+};
 
 export default AspectRatio;
