+++ src/components/Typography/index.tsx
@@ -2,8 +2,13 @@ import React, { FC } from 'react';
 import styled from 'styled-components';
 
 import useTheme from '../../useTheme';
-import { textBase, TextProps } from '../Text';
 import { TypographyAs, TypographyType } from './theme';
+import { TextBaseProps, textBase } from '../../bases';
+import { defaults } from '../../utils/defaults';
+import { typeVariant } from '../../utils/variant';
+import { compose } from '../../utils/baseStyle';
+
+export type TypographyProps = TextBaseProps;
 
 export type ExtendedTypographyProps =
   {
@@ -11,26 +16,30 @@ export type ExtendedTypographyProps =
     t?: TypographyType;
     as?: TypographyAs;
   }
-  & TextProps;
+  & TypographyProps;
 
 const Typography: FC<ExtendedTypographyProps> = ({
-  type: _type,
-  t: _t,
   as: _as,
-  ...props
+  ...p
 }) => {
-  const { typography: { default: _default, types } } = useTheme();
-  const type = _type ?? _t ?? 'p';
+  const { typography } = useTheme();
+  const type = p.type ?? p.t ?? 'p';
 
   const as: TypographyAs = _as ?? (type === 'lead' ? 'p' : type);
 
+  const props = defaults(
+    p,
+    typeVariant(typography, 'p', p),
+    typography.default
+  );
+
   return (
-    <TypographyStyle {..._default} {...types[type]} {...props} as={as} />
+    <TypographyStyle {...props} as={as} />
   );
 };
 
 export default Typography;
 
-const TypographyStyle = styled.p<TextProps>`
-  ${p => textBase(p)}
+const TypographyStyle = styled.p<TypographyProps>`
+  ${compose(textBase)}
 `;
